Internal CBB displays generic data
Creates a dummy gameobject to access all the instance members of Actions and Sensors (the ones implementing IGeneric), then remove the dummyObject. This workflow was created because the members are not accesible directly i.e. you need to create an instance, but since is a monobehaviour, it needs to be attached to a gameObject instead  of just creating a new instance